A bunch of refactoring and adding tests

+++ airborn.web/Models/Aircraft/JsonBindings/PressureAltitudePerformanceProfileNotFoundException.cs
@@ -4,7 +4,7 @@ namespace Airborn.web.Models
 {
     public class PressureAltitudePerformanceProfileNotFoundException : JsonPerformanceProfileNotFoundExceptionBase
     {
-        public PressureAltitudePerformanceProfileNotFoundException(int pressureAltitude) :
+        public PressureAltitudePerformanceProfileNotFoundException(decimal pressureAltitude) :
             base(String.Format($"No performance profile found for pressure altitude: {pressureAltitude} ft"))
         {
             PressureAltitude = pressureAltitude;
